byte, not short

+++ app/src/main/java/us/egeler/matt/obc/mapDataManager/osmXmlReader/OsmXmlReader.java
@@ -250,7 +250,7 @@ public class OsmXmlReader {
                     e.user = attrVal;
                     break;
                 case "layer":
-                    e.layer = Short.parseShort(attrVal);
+                    e.layer = Byte.parseByte(attrVal);
                     break;
             }
         }
